MINOR: new function my_realloc2 = realloc + free upon failure
When realloc fails to allocate memory, the original pointer is not
freed. Sometime people override the original pointer with the pointer
returned by realloc which is NULL in case of failure. This results
in a memory leak because the memory pointed by the original pointer
cannot be freed.

(cherry picked from commit 2eae3a0497f01960dd7f4ffc638f11fc36aa5991)
[wt: backported just for next patch]